Only rethrow error from resolving when it is an unknown error code.
When operating under PnP, if the thing being required does not exist as
a dependency or devDependency the PnP system will throw an error
(`UNDECLARED_DEPENDENCY`). We were previously checking for the error
properly, but after handling it we rethrew (which breaks consumers like
ember-cli-version-checker unless we expect them to wrap in a try/catch).

This fix ensures that we return `null` in this case (which matches the
behavior without PnP when a package is missing).